src/backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
opencv-python==4.5.3.56
torch==1.9.0
Pillow==8.3.1
roboflow==1.0.5
